requirements.txt
absl-py==2.0.0
apex==0.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
contextlib2==21.6.0
contourpy==1.1.1
cycler==0.12.1
fonttools==4.47.2
google-auth==2.26.2
google-auth-oauthlib==1.0.0
grpcio==1.60.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
joblib==1.3.2
kiwisolver==1.4.5
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.7.4
ml-collections==0.1.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==23.2
pandas==2.0.3
pillow==10.2.0
pip==23.3.1
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
setuptools==59.5.0
six==1.16.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
torch==1.10.1+cu111
torchaudio==0.10.1+cu111
torchsummary==1.5.1
torchvision==0.11.2+cu111
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
wheel==0.41.2
zipp==3.17.0
